+++ appenbuf.h
@@ -52,7 +52,7 @@
 	
 	#define ABUF_INIT {NULL,0}
 	
-	void abAppend( struct abuf *ab, const char *s, int len );
+	void abAppend( struct abuf *ab, const char *s, size_t len );
 	
 	void abFree( struct abuf *ab );
